// src/lib/schemas/activityType.ts
enum ActivityType {
  ProjectCreated = "PROJECT_CREATED",
  ProjectUpdated = "PROJECT_UPDATED",
  ProjectDeleted = "PROJECT_DELETED",
  TaskCreated = "TASK_CREATED",
  TaskUpdated = "TASK_UPDATED",
  TaskDeleted = "TASK_DELETED",
  SprintCreated = "SPRINT_CREATED",
  SprintUpdated = "SPRINT_UPDATED",
  SprintDeleted = "SPRINT_DELETED",
  CommentCreated = "COMMENT_CREATED",
  CommentUpdated = "COMMENT_UPDATED",
  CommentDeleted = "COMMENT_DELETED",
  TimeTrackCreated = "TIME_TRACK_CREATED",
  TimeTrackUpdated = "TIME_TRACK_UPDATED",
  TimeTrackDeleted = "TIME_TRACK_DELETED",
}

export { ActivityType };